Struct nagiosplugin::PartialOrdMetric

pub struct PartialOrdMetric<T> where
    T: PartialOrd + ToPerfString + Clone,  { /* fields omitted */ }

A PartialOrdMetric is a metric which will automatically calculate the State based on the given value and warning and/or critical value.

It doesn't matter if you provide warning or critical or both of none of these. Even though you should choose SimpleMetric if you aren't providing any warning or critical value.

The state function of the implemented Metric trait will always be one of: Ok, Warning, Critical

let metric = PartialOrdMetric::new("test", 15, Some(15), Some(30), None, None, false);
assert_eq!(metric.state(), Some(State::Warning));
assert_eq!(metric.value(), 15);

Methods

impl<T> PartialOrdMetric<T> where
T: PartialOrd + ToPerfString + Clone,

pub fn new(
name: &str,
value: T,
warning: Option<T>,
critical: Option<T>,
min: Option<T>,
max: Option<T>,
lower_is_critical: bool
) -> Self

Creates a new PartialOrdMetric from the given values.

In debug builds this will panic if you pass incorrect values for warning and critical.