Trait nagiosplugin::ToPerfString

pub trait ToPerfString {
    fn to_perf_string(&self) -> String;
}

The purpose of ToPerfString is only so one can define custom representations of custom types without using the ToString trait so we don't interfere with that.

Also used internally for generation of the final output.

It's already implemented for some basic types.

Required Methods

fn to_perf_string(&self) -> String